+++ app/models/user.rb
@@ -11,9 +11,12 @@ class User < ApplicationRecord
     # validates :link_site
     # validates :link_fb
     # validates :link_tt
+
     has_many :follows
     has_many :orgs, through: :follows
 
+    has_many :subscribeds
+    has_many :events, :through => :subscribeds
 
     has_secure_password
     validates :password, presence: true, length: { minimum: 6 }
@@ -24,5 +27,4 @@ class User < ApplicationRecord
                 BCrypt::Engine.cost
                 BCrypt::Password.create(string, cost: cost)
   end
-
 end