+++ db/schema.rb
@@ -36,6 +36,15 @@ ActiveRecord::Schema.define(version: 2019_05_16_022128) do
     t.index ["user_id"], name: "index_follows_on_user_id"
   end
 
+  create_table "hosts", force: :cascade do |t|
+    t.bigint "org_id"
+    t.bigint "event_id"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.index ["event_id"], name: "index_hosts_on_event_id"
+    t.index ["org_id"], name: "index_hosts_on_org_id"
+  end
+
   create_table "orgs", force: :cascade do |t|
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
@@ -49,6 +58,15 @@ ActiveRecord::Schema.define(version: 2019_05_16_022128) do
     t.string "link_site"
     t.string "link_fb"
     t.string "link_tt"
+  end
+
+  create_table "subscribeds", force: :cascade do |t|
+    t.bigint "user_id"
+    t.bigint "event_id"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.index ["event_id"], name: "index_subscribeds_on_event_id"
+    t.index ["user_id"], name: "index_subscribeds_on_user_id"
   end
 
   create_table "users", force: :cascade do |t|
@@ -65,4 +83,8 @@ ActiveRecord::Schema.define(version: 2019_05_16_022128) do
     t.string "password_digest"
   end
 
+  add_foreign_key "hosts", "events"
+  add_foreign_key "hosts", "orgs"
+  add_foreign_key "subscribeds", "events"
+  add_foreign_key "subscribeds", "users"
 end